Flask Integration
A very large update that alows flask to be implemented. This meant changing many links and locations in the html documents, and a few additions to the python program that runs the html. I also tidied up a lot of code and remove some things to make Flask implementation easier.